NOD-256: Impostare log fdrnodo come quelli di ndp
